requirements_visualization.txt
# Requirements for MongoDB Data Visualization and ML
pymongo
python-dotenv
matplotlib
numpy
seaborn
scipy
scikit-learn
joblib
pandas
